// TM8/tm8-game-be-main/src/common/exceptions/custom.exception.enum.ts
export enum ExceptionName {
  INCORRECT_SIGNIN = 'INCORRECT_SIGNIN',
  VALIDATION_FAILED = 'VALIDATION_FAILED',
  INTERNAL_EXCEPTION = 'INTERNAL_EXCEPTION',
  FORBIDDEN = 'FORBIDDEN',
  NOT_FOUND = 'NOT_FOUND',
  ALREADY_EXISTS = 'ALREADY_EXISTS',
  CONFLICT = 'CONFLICT',
  EXTERNAL_SERVICE_EXCEPTION = 'EXTERNAL_SERVICE_EXCEPTION',
  REFRESH_TOKEN = 'REFRESH_TOKEN',
  UNAUTHORIZED = 'UNAUTHORIZED',
  BAD_REQUEST = 'BAD_REQUEST',
  THROTTLER = 'THROTTLER',
  INVALID_USER_STATUS = 'INVALID_USER_STATUS',
}